Refactor inference functions. Add draw text. Allow list of class paths.
