+++ official/vision/beta/serving/run_lib.py
@@ -1,13 +1,17 @@
 r"""Vision models run inference utility function."""
 
-from typing import Callable, Optional
+from typing import Callable, Optional, List, Mapping
 import os
 import glob
+import json
+import random
+import colorsys
 
 from absl import flags
+import cv2
+import numpy as np
 import tensorflow as tf
 
-# from official.common import registry_imports  # pylint: disable=unused-import
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.vision.beta import configs
@@ -32,7 +36,8 @@ def define_flags():
   # optional flags, supplied as kwargs
   flags.DEFINE_boolean('visualise', None, '(Segmentation) Flag to visualise mask.')
   flags.DEFINE_boolean('stitch_original', None, '(Segmentation) Flag to stitch image at the side.')
-  flags.DEFINE_string('class_names_path', None, '(Yolo) Txt file with class names.')
+  flags.DEFINE_string('class_names_path', None, '(Yolo/Cls) Csv of paths to txt' + \
+    ' files with class names.')
 
 
 def inference_dataset(image_path_glob: str,
@@ -133,3 +138,97 @@ def get_export_module(experiment: str,
         type(params.task)))
 
   return export_module
+
+
+def read_class_names(class_names_path):
+  """Reads class names from text file.
+  Supports .txt and .json.
+  
+  Args:
+    class_names_path: `str`, path to json/txt file containing classes. 
+      Text file should contain one class name per line.
+      Json file should contain only one dictionary, `Mapping[int, str]`
+  """
+  
+  names = {}
+  if class_names_path.endswith('.txt'):
+    with open(class_names_path, 'r') as data:
+      for idx, name in enumerate(data):
+        names[idx] = name.strip('\n')
+  
+  elif class_names_path.endswith('.json'):
+    with open(class_names_path) as f:
+      names = json.load(f)
+    if type(list(names.keys())[0]) == str and type(list(names.values())[0]) == int:
+      names = dict((v,k) for k,v in names.items())
+
+  else:
+    raise NotImplementedError('File type is not .txt or .json, path %s' %class_names_path)
+
+  assert type(list(names.keys())[0]) == int, 'Loaded dict %s has wrong key type %s' %(
+    class_names_path, type(list(names.keys()[0])))
+  assert type(list(names.values())[0]) == str, 'Loaded dict %s has wrong value type %s' %(
+    class_names_path, type(list(names.values()[0])))
+
+  return names
+
+
+def draw_bbox(image: np.array, 
+              bboxes: np.array,
+              scores: np.array,
+              classes: np.array,
+              num_bboxes: np.array,
+              class_names: Mapping[int, str], 
+              show_label: bool = True):
+  num_classes = len(class_names)
+  image_h, image_w, _ = image.shape
+  hsv_tuples = [(1.0 * x / num_classes, 1., 1.) for x in range(num_classes)]
+  colors = list(map(lambda x: colorsys.hsv_to_rgb(*x), hsv_tuples))
+  colors = list(map(lambda x: (int(x[0] * 255), int(x[1] * 255), int(x[2] * 255)), colors))
+
+  random.seed(0)
+  random.shuffle(colors)
+  random.seed(None)
+
+  for i in range(num_bboxes[0]):
+    if int(classes[0][i]) < 0 or int(classes[0][i]) > num_classes: continue
+    coor = bboxes[0][i] * [image_h, image_w, image_h, image_w]
+    coor = coor.astype(np.int32)
+
+    fontScale = 0.5
+    score = scores[0][i]
+    class_ind = int(classes[0][i])
+    bbox_color = colors[class_ind]
+    bbox_thick = int(0.6 * (image_h + image_w) / 600)
+    c1, c2 = (coor[1], coor[0]), (coor[3], coor[2])
+    cv2.rectangle(image, c1, c2, bbox_color, bbox_thick)
+
+    if show_label:
+      bbox_mess = '%s: %.2f' % (class_names[class_ind], score)
+      t_size = cv2.getTextSize(bbox_mess, 0, fontScale, thickness=bbox_thick // 2)[0]
+      c3 = (c1[0] + t_size[0], c1[1] - t_size[1] - 3)
+      cv2.rectangle(image, c1, c3, bbox_color, -1) #filled
+
+      cv2.putText(image, bbox_mess, (c1[0], c1[1] - 2), cv2.FONT_HERSHEY_SIMPLEX,
+            fontScale, (0, 0, 0), bbox_thick // 2, lineType=cv2.LINE_AA)
+  return image
+
+
+def draw_text(image: np.array, 
+              text_list: List[str],
+              spacing: int = 50):
+    """ Writes list of messages on image, separated by newline 
+    
+    Args:
+      image: `np.array`, of shape (height, width, 3), RGB image
+      text_list: `List[str]`, list of texts to write on each line
+      spacing: `int`, spacing in pixels
+    """
+
+    pos = spacing
+    for text in text_list:
+        image = cv2.putText(image, text,
+                            (spacing, pos), cv2.FONT_HERSHEY_SIMPLEX, 0.6, (255, 0, 0), 2)
+        pos += spacing
+
+    return image
